separate out state from node
